Catch extra results coming in the <div class="srg"> element
This little patch helps to catch results that come with an intermediate div tag between the <div class="rso"> and the <div class="g"> results. Sometimes Google spits them out like this, especially where there are adwords or sitelinks involved.

When this "anomalous" intermediate tag is detected, it's children nodes are returned for parsing as simple results.